Remove comment from LICENSE.md
The recommendation is not to modify/add to the MIT license (and probably not other licenses). It also breaks GitHub's license integration that will show a nice summary of the license as a header. See e.g. https://github.com/andreasnoack/TSVD.jl/blob/master/LICENSE.md. The deleted information is still in the beginning of `README.md`.