fixed compat entry for GZIP
my previous commit, 476e635, allowed compatibility with GZip 0.6 versions, switched back to 0.5.1 (equivalent to ^0.5.1, as carats are default), which states that there is compatibility with anything greater than or equal to 0.5.1, and anything less than 0.6.0, i.e. [0.5.1, 0.6.0).

I followed precedent from the compatibility entry for IterativeSolvers, 0.7, which is equivalent to ^0.7 and specifies a range of [0.7.0, 0.8.0). See https://julialang.github.io/Pkg.jl/v1/compatibility/index.html.

IterativeSolvers was set to 0.7.1 before a different commit (614b86c), equivalent to [0.7.1, 0.8.0), cc'ing @ranjanan in case the lower compatibility limit actually needs to be 0.7.1 due to some important patch (same goes for AlgebraicMultigrid, which was at 0.2.1).

Also cc @ViralBShah

Maybe we should open an issue if there needs to be more discussion.